+++ backend/db/competition/playerDatabase.js
@@ -6,6 +6,7 @@ let ObjectID = require( 'mongodb' ).ObjectID;
 
 
 exports.getPlayerById = async ( id ) => {
+	if ( !ObjectID.isValid( id ) ) throw { code: 422, message: "Invalid player id" };
 	let result = ( await playerCursor.findOne( { _id: ObjectID( id.toString() ) } ) );
 	return result;
 };
@@ -33,12 +34,19 @@ exports.createPlayer = async ( player ) => {
 };
 
 exports.updatePlayer = async ( id, player ) => {
+	if ( !ObjectID.isValid( id ) ) throw { code: 422, message: "Invalid player id" };
 	let result = ( await playerCursor.findOneAndUpdate( { _id: ObjectID( id.toString() ) }, { $set: player }, { returnOriginal: false } ) );
 	return result.value;
 };
 
 exports.deletePlayer = async ( id ) => {
+	if ( !ObjectID.isValid( id ) ) throw { code: 422, message: "Invalid player id" };
 	let result = ( await playerCursor.deleteOne( { _id: ObjectID( id.toString() ) } ) );
 	return ( result.result.n === 1 && result.result.ok === 1 );
 };
 
+
+exports.countTeamPlayers = async ( teamID ) => {
+	if ( !ObjectID.isValid( teamID ) ) throw { code: 422, message: "Invalid team id" };
+	return ( await ( await playerCursor.find( { "competitions.teamID": teamID } ) ).count() );
+};
